Extract a helper for writing schema files in tests
