+++ packages/flutter_tools/test/flutter_manifest_test.dart
@@ -12,6 +12,7 @@ import 'package:flutter_tools/src/flutter_manifest.dart';
 
 import 'src/common.dart';
 import 'src/context.dart';
+import 'src/schema.dart';
 
 void main() {
   setUpAll(() {
@@ -515,23 +516,11 @@ flutter:
       expect(flutterManifest.isEmpty, false);
     }
 
-    void writeSchemaFile(FileSystem filesystem, String schemaData) {
-      final String schemaPath = buildSchemaPath(filesystem);
-      final File schemaFile = filesystem.file(schemaPath);
-
-      final String schemaDir = buildSchemaDir(filesystem);
-
-      filesystem.directory(schemaDir).createSync(recursive: true);
-      filesystem.file(schemaFile).writeAsStringSync(schemaData);
-    }
-
     void testUsingContextAndFs(String description, FileSystem filesystem,
         dynamic testMethod()) {
-      const String schemaData = '{}';
-
       testUsingContext(description,
               () async {
-            writeSchemaFile( filesystem, schemaData);
+            writeBasicSchemaFile(filesystem);
             testMethod();
       },
           overrides: <Type, Generator>{
